+++ RedditClone.Infrastructure/Persistence/Repositories/PostRepository.cs
@@ -1,36 +1,53 @@
-namespace RedditClone.Infrastructure.Persistence;
+namespace RedditClone.Infrastructure.Persistence.Repositories;
 
-using System.Net;
+using ErrorOr;
+using Serilog;
 using Microsoft.EntityFrameworkCore;
-using RedditClone.Application.Common.Errors;
-using RedditClone.Application.Persistence;
-using RedditClone.Domain.Common.ValueObjects;
-using RedditClone.Domain.CommunityAggregate.ValueObjects;
+using RedditClone.Domain.Common.Errors;
 using RedditClone.Domain.PostAggregate;
+using RedditClone.Domain.Common.ValueObjects;
 using RedditClone.Domain.PostAggregate.Entities;
 using RedditClone.Domain.PostAggregate.ValueObjects;
 using RedditClone.Domain.UserAggregate.ValueObjects;
+using RedditClone.Domain.CommunityAggregate.ValueObjects;
+using RedditClone.Application.Common.Interfaces.Persistence;
 
-public class PostRepository : IPostRepository
+public class PostRepository(RedditCloneDbContext dbContext)
+    : IPostRepository
 {
-    private readonly RedditCloneDbContext _dbContext;
+    private readonly RedditCloneDbContext _dbContext = dbContext;
 
-    public PostRepository(RedditCloneDbContext dbContext)
+    public ErrorOr<Post> GetPostById(PostId postId)
     {
-        _dbContext = dbContext;
-    }
-    public Post GetPostById(PostId postId)
-    {
-        Post post = _dbContext.Posts.FirstOrDefault(p => p.Id == postId)
-        ?? throw new HttpCustomException(
-        HttpStatusCode.NotFound, "Post not found");
+        Post? post = _dbContext.Posts.FirstOrDefault(p => p.Id == postId);
+
+        if (post is null)
+        {
+            Error error = Errors.Posts.PostNotFound;
+
+            Log.Error(
+                "{@Code}, {@Description}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
 
         return post;
     }
 
+    public List<Votes> GetVotesListsByPostId(PostId postId)
+    {
+        List<Votes> votes = _dbContext.Posts.FirstOrDefault(p => p.Id == postId)!.Votes.ToList();
+
+        return votes;
+    }
+
     public List<Post> GetPostListByUser(UserId userId)
     {
-        List<Post> posts = _dbContext.Posts.Where(p => p.UserId == userId).ToList();
+        List<Post> posts = _dbContext.Posts.Where(p => p.UserId == userId)
+        .Include(p => p.Votes)
+        .ToList();
 
         return posts;
     }
@@ -42,48 +59,72 @@ public class PostRepository : IPostRepository
         return posts;
     }
 
-    public async void Add(Post post)
+    public void Add(Post post)
     {
         _dbContext.Posts.Add(post);
-
-        await _dbContext.SaveChangesAsync();
     }
 
-    public Post UpdatePostById(PostId id, UserId userId, string title, string content)
+    public ErrorOr<Post> UpdatePostById(PostId id, UserId userId, string title, string content)
     {
-        Post post = _dbContext.Posts.SingleOrDefault(p => p.Id == id && p.UserId == userId)
-            ?? throw new HttpCustomException(
-            HttpStatusCode.NotFound, "Post not found on you posts");
+        Post? post = _dbContext.Posts.SingleOrDefault(p => p.Id == id && p.UserId == userId);
+
+        if (post is null)
+        {
+            Error error = Errors.Posts.PostNotFound;
+
+            Log.Error(
+                "{@Code}, {@Description}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
 
         post.UpdatePost(title, content);
 
         _dbContext.Posts.Update(post);
 
-        _dbContext.Entry(post).State = EntityState.Modified;
-
-        _dbContext.SaveChanges();
-
         return post;
     }
 
-    public void DeletePostById(PostId id, UserId userId)
+    public ErrorOr<bool> DeletePostById(PostId id, UserId userId)
     {
-        Post post = _dbContext.Posts.SingleOrDefault(p => p.Id == id && p.UserId == userId)
-            ?? throw new HttpCustomException(
-            HttpStatusCode.NotFound, "Post not found on you posts");;
+        Post? post = _dbContext.Posts.SingleOrDefault(p => p.Id == id && p.UserId == userId);
+
+        if (post == null)
+        {
+            Error error = Errors.Posts.PostNotFound;
+
+            Log.Error(
+                "{@Code}, {@Description}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
 
         _dbContext.Posts.Remove(post);
 
-        _dbContext.SaveChanges();
+        return true;
     }
 
-    public void AddPostVote(PostId id, UserId userId, bool isVoted)
+    public ErrorOr<bool> AddPostVote(PostId id, UserId userId, bool isVoted)
     {
-        Post postVote = _dbContext.Posts
+        Post? postVote = _dbContext.Posts
             .Include(p => p.Votes)
-            .SingleOrDefault(p => p.Id == id)
-            ?? throw new HttpCustomException(
-            HttpStatusCode.NotFound, "Post not found");;
+            .SingleOrDefault(p => p.Id == id);
+
+        if (postVote == null)
+        {
+            Error error = Errors.Posts.PostNotFound;
+
+            Log.Error(
+                "{@Code}, {@Description}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
 
         var vote = Votes.Create(id, userId, isVoted);
 
@@ -91,40 +132,56 @@ public class PostRepository : IPostRepository
 
         _dbContext.Posts.Update(postVote);
 
-        _dbContext.Entry(postVote).State = EntityState.Modified;
-
-        _dbContext.SaveChanges();
+        return true;
     }
 
-    public void UpdatePostVoteById(PostId id, VoteId voteId, UserId userId, bool isVoted)
+    public ErrorOr<bool> UpdatePostVoteById(PostId id, VoteId voteId, UserId userId, bool isVoted)
     {
-        Post postVote = _dbContext.Posts
+        Post? postVote = _dbContext.Posts
             .Include(p => p.Votes)
             .Where(p => p.Votes.Any(pv => pv.Id == voteId && pv.UserId == userId))
-            .SingleOrDefault(p => p.Id == id)
-            ?? throw new HttpCustomException(
-            HttpStatusCode.NotFound, "Vote not found on that post");
+            .SingleOrDefault(p => p.Id == id);
+
+        if (postVote == null)
+        {
+            Error error = Errors.Posts.VoteNotFound;
+
+            Log.Error(
+                "{@Code}, {@Description}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
 
         postVote.UpdateVote(voteId, isVoted);
 
-        _dbContext.Entry(postVote).State = EntityState.Modified;
-
-        _dbContext.SaveChanges();
+        return true;
     }
 
-    public void DeletePostVoteById(PostId id, VoteId voteId, UserId userId)
+    public ErrorOr<bool> DeletePostVoteById(PostId id, VoteId voteId, UserId userId)
     {
-        Post postVote = _dbContext.Posts
+        Post? postVote = _dbContext.Posts
             .Include(p => p.Votes)
             .Where(p => p.Votes.Any(pv => pv.Id == voteId && pv.UserId == userId))
-            .SingleOrDefault(p => p.Id == id)
-            ?? throw new HttpCustomException(
-            HttpStatusCode.NotFound, "Vote not found on that post");
+            .SingleOrDefault(p => p.Id == id);
+
+        if(postVote == null)
+        {
+            Error error = Errors.Posts.VoteNotFound;
+
+            Log.Error(
+                "{@Code}, {@Description}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
+
+        _dbContext.Posts.Remove(postVote);;
 
         postVote.RemoveVote(voteId);
 
-        _dbContext.Entry(postVote).State = EntityState.Modified;
-
-        _dbContext.SaveChanges();
+        return true;
     }
 }
